move CodeDB and StateDB
