+++ Cargo.toml
@@ -62,7 +62,7 @@ strum_macros = "0.25"
 subtle = "2.4"
 tokio = { version = "1.13", features = ["macros", "rt-multi-thread"] }
 url = "2.2"
-revm-precompile = { git = "https://github.com/scroll-tech/revm", branch = "scroll-fix" }
+revm-precompile = { git = "https://github.com/scroll-tech/revm", rev = "b57e7d9" } # v35
 c-kzg = "1.0.0"
 
 [patch.crates-io]
