+++ eth-types/Cargo.toml
@@ -25,11 +25,12 @@ strum_macros.workspace = true
 strum.workspace = true
 hash-circuit.workspace = true
 base64.workspace = true
+revm-precompile.workspace = true
 
 [features]
 default = ["warn-unimplemented"]
 warn-unimplemented = []
-scroll = []
+scroll = ["revm-precompile/scroll"]
 
 # trace heap allocation related feature switches
 enable-stack = []
