+++ eth-types/src/lib.rs
@@ -28,6 +28,8 @@ pub mod evm_types;
 pub mod geth_types;
 pub mod l2_types;
 pub mod sign_types;
+pub mod state_db;
+pub mod utils;
 
 use crate::evm_types::{Gas, GasCost, OpcodeId, ProgramCounter};
 pub use bytecode::Bytecode;
