+++ zkevm-circuits/src/bytecode_circuit/circuit.rs
@@ -8,8 +8,9 @@ use crate::{
     util::{get_push_size, Challenges, Expr, SubCircuit, SubCircuitConfig},
     witness,
 };
-use bus_mapping::{state_db::EMPTY_CODE_HASH_LE, util::POSEIDON_CODE_HASH_EMPTY};
-use eth_types::{Field, ToLittleEndian, ToScalar, ToWord};
+use eth_types::{
+    state_db::EMPTY_CODE_HASH_LE, Field, ToLittleEndian, ToScalar, ToWord, POSEIDON_CODE_HASH_EMPTY,
+};
 use gadgets::is_zero::{IsZeroChip, IsZeroConfig, IsZeroInstruction};
 use halo2_proofs::{
     circuit::{Layouter, Region, Value},
