+++ zkevm-circuits/src/evm_circuit/execution/begin_tx.rs
@@ -29,11 +29,8 @@ use crate::{
     },
 };
 use array_init::array_init;
-use bus_mapping::{
-    circuit_input_builder::CopyDataType,
-    precompile::{is_precompiled, PrecompileCalls},
-};
-use eth_types::{Address, Field, ToLittleEndian, ToScalar, U256};
+use bus_mapping::{circuit_input_builder::CopyDataType, precompile::PrecompileCalls};
+use eth_types::{utils::is_precompiled, Address, Field, ToLittleEndian, ToScalar, U256};
 use ethers_core::utils::{get_contract_address, keccak256, rlp::RlpStream};
 use gadgets::util::{expr_from_bytes, not, select, Expr};
 use halo2_proofs::{circuit::Value, plonk::Error};
