+++ zkevm-circuits/src/evm_circuit/execution/callop.rs
@@ -24,12 +24,11 @@ use crate::{
     util::Expr,
 };
 use bus_mapping::{
-    circuit_input_builder::CopyDataType,
-    evm::OpcodeId,
-    precompile::{is_precompiled, PrecompileCalls},
+    circuit_input_builder::CopyDataType, evm::OpcodeId, precompile::PrecompileCalls,
 };
 use eth_types::{
     evm_types::{memory::MemoryWordRange, GAS_STIPEND_CALL_WITH_VALUE},
+    utils::is_precompiled,
     Field, ToAddress, ToBigEndian, ToLittleEndian, ToScalar, U256,
 };
 use halo2_proofs::{circuit::Value, plonk::Error};
