+++ zkevm-circuits/src/evm_circuit/execution/return_revert.rs
@@ -20,9 +20,10 @@ use crate::{
     table::{AccountFieldTag, CallContextFieldTag},
     util::Expr,
 };
-use bus_mapping::{circuit_input_builder::CopyDataType, state_db::CodeDB};
+use bus_mapping::circuit_input_builder::CopyDataType;
 use eth_types::{
     evm_types::{GasCost, OpcodeId},
+    state_db::CodeDB,
     Field, ToScalar, U256,
 };
 use ethers_core::utils::keccak256;
