+++ zkevm-circuits/src/evm_circuit/util/constraint_builder.rs
@@ -11,11 +11,8 @@ use crate::{
     },
     util::{build_tx_log_expression, Challenges, Expr},
 };
-use bus_mapping::{
-    state_db::EMPTY_CODE_HASH_LE,
-    util::{KECCAK_CODE_HASH_EMPTY, POSEIDON_CODE_HASH_EMPTY},
-};
-use eth_types::{Field, ToLittleEndian, ToScalar, ToWord};
+use bus_mapping::util::{KECCAK_CODE_HASH_EMPTY, POSEIDON_CODE_HASH_EMPTY};
+use eth_types::{state_db::EMPTY_CODE_HASH_LE, Field, ToLittleEndian, ToScalar, ToWord};
 use gadgets::util::{and, not};
 use halo2_proofs::{
     circuit::Value,
