+++ zkevm-circuits/src/evm_circuit/util/math_gadget/rlp.rs
@@ -422,8 +422,7 @@ impl<F: Field, const IS_CREATE2: bool> ContractCreateGadget<F, IS_CREATE2> {
 #[cfg(test)]
 mod test {
     use super::{super::test_util::*, ContractCreateGadget};
-    use bus_mapping::state_db::CodeDB;
-    use eth_types::{Field, ToAddress, ToLittleEndian, ToWord, Word};
+    use eth_types::{state_db::CodeDB, Field, ToAddress, ToLittleEndian, ToWord, Word};
     use ethers_core::utils::keccak256;
     use gadgets::util::{not, Expr};
     use halo2_proofs::halo2curves::bn256::Fr;
