+++ zkevm-circuits/src/witness/block.rs
@@ -463,7 +463,7 @@ impl From<&circuit_input_builder::Block> for BlockContexts {
 /// Convert a block struct in bus-mapping to a witness block used in circuits
 pub fn block_convert<F: Field>(
     block: &circuit_input_builder::Block,
-    code_db: &bus_mapping::state_db::CodeDB,
+    code_db: &eth_types::state_db::CodeDB,
 ) -> Result<Block<F>, Error> {
     let rws = RwMap::from(&block.container);
     rws.check_value()?;
@@ -581,7 +581,7 @@ pub fn block_convert<F: Field>(
 /// Convert a block struct in bus-mapping to a witness block used in circuits
 pub fn block_convert_with_l1_queue_index<F: Field>(
     block: &circuit_input_builder::Block,
-    code_db: &bus_mapping::state_db::CodeDB,
+    code_db: &eth_types::state_db::CodeDB,
     start_l1_queue_index: u64,
 ) -> Result<Block<F>, Error> {
     let mut block = block.clone();
